Armature speed ups, Part I --------------------------
Major speed up for armatures during times when you aren't
posing a figure.

Background: the calculation of poses generated by actions and the
calculation of displists were getting somewhat out of sync.
This was being remedied by 'clearing the constraint done flag'
of the pose channels and recalculating the displists every time
the 3d view was redrawn, making life slow and unpleasant.
Commenting out the code that was doing this, then reinserting
the 'clearing the constraint done flag' at the right times
made things a bit more perky.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@1793 954f8c5b-7b00-dc11-b283-0030488c597c