Rewind of change in MCol structure.
Kester: since this is in DNA, it backwards converts all R into B now,
which isn't the bug. On big endian systems (OSX) there engine colors
are OK, which indicates the conversion of MCol goes wrong on the
engine side, by using some endian-sensative approach.

The main confusement is, that the MCol struct isn't in use for read
or write in Blender, only for a save in files. The contents of that
struct is wrong, yes. See note in vpaint.c top.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@2553 954f8c5b-7b00-dc11-b283-0030488c597c