Fixing something else broken broke: The fix for "no action == crash" prevented rvk editing to work in the action window.
Please not that "if (act) {" (Does the object have an action? Currently
only true for some armatures) and "if (key) {" (does the object have
rvk's keys? Currently only true for meshes) are never both true at the
same time so in the following snippet, the function foo_key() never gets
executed:

if (act) {
  if (key) {
    foo_key();
  else
    foo_bar();
  }
}

Better is this:

if (key) {
  foo_key();
else if (act) {
  foo_bar();
}


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@2632 954f8c5b-7b00-dc11-b283-0030488c597c