Enabled trial for 'global undo'. This now only has undo steps for - transform (grab, rot, scale, etc) - all button commands, including menus
I sacrificed for now the UKEY in the 3d window for it. Shift+U does a redo.
(Only in 3d window)

What this system does is saving files in the temp directory (user pref).
The filenames are cycled around (32 in total now).

This commit will follow shortly with a userpref for it, not to frustrate
people who want to work in normal fashion with blender.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@3034 954f8c5b-7b00-dc11-b283-0030488c597c