fix numerous memory leaks in the math types. fixed bug:  #1633  Memory leak in M_Mathutils_Vector
The math types ( matrix, vector, quad ) now make
copies of data passed to them rather than holding
a pointer to memory that cannot be freed, or that
may go away unexpectedly.

This also clarifies the problem of who is responsible
for freeing memory allocations.

Pre-checkin files are tagged mem_leak-1 in case this
breaks something.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@3239 954f8c5b-7b00-dc11-b283-0030488c597c