Martin DeMello's patch to add dxf import code for: CIRCLE, ARC, ELLIPSE and LWPOLYLINE.
I made some modifications to it to clean it up a bit.

I entend to do some more dxf stuff in  the near future but figured
this is ready to go for now, and trying to clean up a bunch of small stuff
I've had sitting around for awhile. ;)

Kent


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@3269 954f8c5b-7b00-dc11-b283-0030488c597c