Fix for #1549
This was something users found since tracing got into blender, having
sometimes small 'dots' or bright pixels or missing reflection rays in
an image. Thanks to the very simple sample file I could disect it...
it appeared to be an incomplete check for all numerical exceptions when
you traverse the octree nodes. Very technical, but clear comments are in
the code to explain ;)


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@3270 954f8c5b-7b00-dc11-b283-0030488c597c