Patch provided by Yann Vernier
This patch should make gettext return unicode, which is what freetype
expects (also explicitly set in the same file). It makes translations
that support non-ascii *and* non-unicode encodings, such as latin 1 for
western european languages, draw correctly. The problem only occurs for
platforms where those encodings are default, i.e. in Linux if
/etc/locale.gen looks like this:
sv_SE ISO-8859-15
If it were to read UTF-8 this patch doesn't change the behaviour at all.
This is probably the case for Mac OS X and perhaps Windows.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@3511 954f8c5b-7b00-dc11-b283-0030488c597c