Fix for 2098
The directory where the last file was saved (in .Blog) wasnt added to the
history menu in the FileWindow when using CTRL+O for example. For
convenience, it now alwyas adds this directory to the end of default
history menu, so you can always jump back while browsing in filewindow.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@3727 954f8c5b-7b00-dc11-b283-0030488c597c