Optimising transformations in PET mode.
Sorting elements in TransData array according to their distance. That means we can always stop loops when it finds the first element that doesn't apply. That makes PET factor calculation and actual transformations faster since it stops at the first element outside of the area.

The Sorting algo is a no swap version of quicksort using the start of each subset as the pivot (random would be better and easy to implement. Need 3 additional lines of code really).


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@3955 954f8c5b-7b00-dc11-b283-0030488c597c