add files for Daniel Dunbar's new CCGSUBSURFLIB work to SCons build. This commit *only* adds the new files.
Since this is an experimental feature, it is NOT enabled by default.

At the moment, you can turn it on by editing the toplevel SConstruct
file.  NOTE:  this is a hack.  it will be done properly later!

To enable this experimental feature,
add the following line after the line that starts with 'extra_flags = ['

    extra_flags.append( '-DUSE_CCGSUBSURFLIB' )


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@4014 954f8c5b-7b00-dc11-b283-0030488c597c