Bug fix #2542
Fun, the error is in code committed by Kent Mein, in a patch provided by
Campbell, who also reported the bug. :)
There was a line that set ALL selected objects (on click) to become active,
in succession, in a loop... very bad.

Activating in Oops better be restricted to ctrl+click. Fixed!


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@4420 954f8c5b-7b00-dc11-b283-0030488c597c