Back to the Future bugfix!
When a new space (window) type gets added, older Blender binaries will
crash on reading files saved with such spaces. This breaks the so nice
upward compility on files...  With addition of SpaceTime I got an occasion
to verify & fix that. If Blender detects a non-existant space it will
degrade it to a 3D window now. Tested on a full 2.34 compile.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@4455 954f8c5b-7b00-dc11-b283-0030488c597c