Changed test for (me->flag&ME_SUBSURF) into                  ((me->flag&ME_SUBSURF) && me->subdiv)
In build_mesh_data(). This because meshes with subdiv==0 were calling a
MakeDisplist() for each redraw otherwise, disabling for example the
posemode "delay deform" feature. (bug #2700)

Daniel; it would help if you would add comments in the code, for example
to understand the difference between;

mesh_get_base_derived()
mesh_get_derived()

and how it relates to

build_mesh_data()

I realize most of the mess is in displist code... and confusing use
of features while editing, posemode, & render. Nevertheless, it's not
clear at first sight which parts of derivedmesh is design, and which
part reflects the hacks to make it work. :)


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@4564 954f8c5b-7b00-dc11-b283-0030488c597c