Fix for bug #2766: Shear shortcut broken
The problem is that the shear shortcut that was originally chosen was
eaten up by the shortcut to switch to the scale blezmo. Changed the
shortcut to CTRL + ALT + SHIFT + S which is painful but seeing as how
shear in object mode has limited use, I don't think it will cause any
real problems. I also updated the menu entry to match.

This was assigned to Martin but he's busy so I did it... and if I did it
wrong... blame him for not fixing it sooner ;)


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@4870 954f8c5b-7b00-dc11-b283-0030488c597c