Fixing (hopefully) a stupid issue with rendering text objects:
cu->curinfo.mat_nr appearently didn't get initialized to 0 when creating
a new curve / text object, thus assigning random material indices to
characters, which crashed when rendering.

I'm still not sure if this was the actual issue, though I'd hope so :)

For old files that crash on rendering, just enter editmode on EACH text
object, select all text, and assign a material to it.

I hope this fixes.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@4972 954f8c5b-7b00-dc11-b283-0030488c597c