- got rid of DerivedMesh.drawMappedEdgeEM function, can be implemented with    drawMappedEdges  - added DerivedMesh.convertToDispListMeshMapped function which converts and    also returns mapping information for use in editmode  - updated DispListMesh derivedmesh to be able to function in editmode  - update mirror modifier to support use as a cage  - update mirror & subsurf modifiers to properly pass mapping information down    modifier stack
It is now possible to have a mesh with mirror/subsurf modifiers where you
can edit with both as cage. Selecting the mirror'd part works, but of course
transform is flipped so it is a bit weird. Not the cleanest code in the
world and I can't say I am really happy with the architecture but it works for
now and supports the existing feature set.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@5034 954f8c5b-7b00-dc11-b283-0030488c597c